COMET-X/requirements.txt
pandas
numpy
torch
torchvision
transformers
opencv-python
scikit-learn
matplotlib
networkx
torch_geometric # Ensure this matches your CUDA version, e.g., torch_geometric==2.5.0+torch2.3.0
requests
Pillow